requirements.txt
fairlearn
scikit-learn
pandas
matplotlib
pytest

azure-identity
azure-mgmt-resource
azure-storage-blob
azure-identity
